+++ aws_lambda_powertools/utilities/kafka/deserializer/base.py
