+++ aws_lambda_powertools/utilities/kafka/deserializer/deserializer.py
@@ -3,11 +3,11 @@ from __future__ import annotations
 import hashlib
 from typing import TYPE_CHECKING, Any
 
-from aws_lambda_powertools.utilities.kafka_consumer.deserializer.default import DefaultDeserializer
-from aws_lambda_powertools.utilities.kafka_consumer.deserializer.json import JsonDeserializer
+from aws_lambda_powertools.utilities.kafka.deserializer.default import DefaultDeserializer
+from aws_lambda_powertools.utilities.kafka.deserializer.json import JsonDeserializer
 
 if TYPE_CHECKING:
-    from aws_lambda_powertools.utilities.kafka_consumer.deserializer.base import DeserializerBase
+    from aws_lambda_powertools.utilities.kafka.deserializer.base import DeserializerBase
 
 # Cache for deserializers
 _deserializer_cache: dict[str, DeserializerBase] = {}
@@ -85,12 +85,12 @@ def get_deserializer(schema_type: str | object, schema_value: Any) -> Deserializ
 
     if schema_type == "AVRO":
         # Import here to avoid dependency if not used
-        from aws_lambda_powertools.utilities.kafka_consumer.deserializer.avro import AvroDeserializer
+        from aws_lambda_powertools.utilities.kafka.deserializer.avro import AvroDeserializer
 
         deserializer = AvroDeserializer(schema_value)
     elif schema_type == "PROTOBUF":
         # Import here to avoid dependency if not used
-        from aws_lambda_powertools.utilities.kafka_consumer.deserializer.protobuf import ProtobufDeserializer
+        from aws_lambda_powertools.utilities.kafka.deserializer.protobuf import ProtobufDeserializer
 
         deserializer = ProtobufDeserializer(schema_value)
     elif schema_type == "JSON":
