+++ aws_lambda_powertools/utilities/kafka/deserializer/json.py
@@ -3,8 +3,8 @@ from __future__ import annotations
 import base64
 import json
 
-from aws_lambda_powertools.utilities.kafka_consumer.deserializer.base import DeserializerBase
-from aws_lambda_powertools.utilities.kafka_consumer.exceptions import KafkaConsumerDeserializationError
+from aws_lambda_powertools.utilities.kafka.deserializer.base import DeserializerBase
+from aws_lambda_powertools.utilities.kafka.exceptions import KafkaConsumerDeserializationError
 
 
 class JsonDeserializer(DeserializerBase):
