+++ aws_lambda_powertools/utilities/kafka/exceptions.py
