+++ aws_lambda_powertools/utilities/kafka/kafka_consumer.py
@@ -3,12 +3,12 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, Any
 
 from aws_lambda_powertools.middleware_factory import lambda_handler_decorator
-from aws_lambda_powertools.utilities.kafka_consumer.consumer_records import ConsumerRecords
+from aws_lambda_powertools.utilities.kafka.consumer_records import ConsumerRecords
 
 if TYPE_CHECKING:
     from collections.abc import Callable
 
-    from aws_lambda_powertools.utilities.kafka_consumer.schema_config import SchemaConfig
+    from aws_lambda_powertools.utilities.kafka.schema_config import SchemaConfig
     from aws_lambda_powertools.utilities.typing import LambdaContext
