+++ aws_lambda_powertools/utilities/kafka/serialization/dataclass.py
@@ -3,8 +3,8 @@ from __future__ import annotations
 from dataclasses import is_dataclass
 from typing import Any, cast
 
-from aws_lambda_powertools.utilities.kafka_consumer.serialization.base import OutputSerializerBase
-from aws_lambda_powertools.utilities.kafka_consumer.serialization.types import T
+from aws_lambda_powertools.utilities.kafka.serialization.base import OutputSerializerBase
+from aws_lambda_powertools.utilities.kafka.serialization.types import T
 
 
 class DataclassOutputSerializer(OutputSerializerBase):
