+++ aws_lambda_powertools/utilities/kafka/serialization/pydantic.py
@@ -4,10 +4,10 @@ from typing import TYPE_CHECKING, Any
 
 from pydantic import TypeAdapter
 
-from aws_lambda_powertools.utilities.kafka_consumer.serialization.base import OutputSerializerBase
+from aws_lambda_powertools.utilities.kafka.serialization.base import OutputSerializerBase
 
 if TYPE_CHECKING:
-    from aws_lambda_powertools.utilities.kafka_consumer.serialization.types import T
+    from aws_lambda_powertools.utilities.kafka.serialization.types import T
 
 
 class PydanticOutputSerializer(OutputSerializerBase):
