--- aws_lambda_powertools/utilities/kafka_consumer/__init__.py
@@ -1,9 +1,0 @@
-from aws_lambda_powertools.utilities.kafka_consumer.consumer_records import ConsumerRecords
-from aws_lambda_powertools.utilities.kafka_consumer.kafka_consumer import kafka_consumer
-from aws_lambda_powertools.utilities.kafka_consumer.schema_config import SchemaConfig
-
-__all__ = [
-    "kafka_consumer",
-    "ConsumerRecords",
-    "SchemaConfig",
-]
