+++ tests/functional/kafka_consumer/_pydantic/test_kafka_consumer.py
@@ -9,12 +9,12 @@ from avro.io import BinaryEncoder, DatumWriter
 from avro.schema import parse as parse_schema
 from pydantic import BaseModel
 
-from aws_lambda_powertools.utilities.kafka_consumer.consumer_records import ConsumerRecords
-from aws_lambda_powertools.utilities.kafka_consumer.exceptions import (
+from aws_lambda_powertools.utilities.kafka.consumer_records import ConsumerRecords
+from aws_lambda_powertools.utilities.kafka.exceptions import (
     KafkaConsumerDeserializationError,
 )
-from aws_lambda_powertools.utilities.kafka_consumer.kafka_consumer import kafka_consumer
-from aws_lambda_powertools.utilities.kafka_consumer.schema_config import SchemaConfig
+from aws_lambda_powertools.utilities.kafka.kafka_consumer import kafka_consumer
+from aws_lambda_powertools.utilities.kafka.schema_config import SchemaConfig
 
 
 @pytest.fixture
