+++ tests/functional/kafka_consumer/required_dependencies/test_kafka_consumer.py
@@ -5,12 +5,12 @@ from dataclasses import dataclass
 
 import pytest
 
-from aws_lambda_powertools.utilities.kafka_consumer.consumer_records import ConsumerRecords
-from aws_lambda_powertools.utilities.kafka_consumer.exceptions import (
+from aws_lambda_powertools.utilities.kafka.consumer_records import ConsumerRecords
+from aws_lambda_powertools.utilities.kafka.exceptions import (
     KafkaConsumerDeserializationError,
 )
-from aws_lambda_powertools.utilities.kafka_consumer.kafka_consumer import kafka_consumer
-from aws_lambda_powertools.utilities.kafka_consumer.schema_config import SchemaConfig
+from aws_lambda_powertools.utilities.kafka.kafka_consumer import kafka_consumer
+from aws_lambda_powertools.utilities.kafka.schema_config import SchemaConfig
 
 
 @pytest.fixture
